+++ docu/examples/example-node-import.mjs
@@ -1,15 +1,14 @@
-const { SOA } = require("../../lib/index"); // run the npm-script "buildTs" to generate js files for this example
-const SCHEMA_URL = "https://raw.githubusercontent.com/semantifyit/schemaorg/main/data/releases/13.0/schemaorg-all-https.jsonld";
-const VOC_OBJ_ZOO = require("../../tests/data/vocabulary-animal.json");
+import { SOA } from "../../lib/index.js"; // run the npm-script "buildTs" to generate js files for this example
 main();
 
 /**
  * example usage of the library within node.js
  */
 async function main() {
-  const mySA = await SOA.create({ vocabularies: [SCHEMA_URL, VOC_OBJ_ZOO] });
+  const mySA = await SOA.create();
   const mySDOUrl = await mySA.constructURLSchemaVocabulary("latest");
   console.log("The latest version is " + mySDOUrl);
+  await mySA.addVocabularies([mySDOUrl]);
   let testClass = mySA.getClass("https://schema.org/Person");
   console.log(mySA.getVocabularies());
   console.log("getIRI() " + testClass.getIRI());
@@ -25,8 +24,6 @@ async function main() {
   const testProp = mySA.getProperty("schema:aspect");
   console.log("isSuperseededBy() " + testProp.isSupersededBy());
 
-  const extVocClass = mySA.getClass("ex:Tiger");
-  console.log(JSON.stringify(extVocClass.toJSON(false, null), null, 2));
   testClass = mySA.getClass("schema:Person");
   console.log(JSON.stringify(testClass.toJSON(false, null), null, 2));
   const testProperty = mySA.getProperty("schema:translationOfWork");