+++ src/SDOAdapter.ts
@@ -24,7 +24,7 @@ import { DataType } from "./DataType";
  * An **SDOAdapter** is an instance of the library itself that holds its own settings and vocabularies (specified by the user). Based on these internal settings and vocabularies the SDOAdapter provides corresponding data through the methods described below (an SDOAdapter can only provide data about a vocabulary, if that vocabulary has been added to the instance). An SDOAdapter instance is created with {@link create | .create()}, have a look at the different settings.
  * ```JS
  * // load the library
- * const SOA = require("schema-org-adapter");
+ * const { SOA } = require("schema-org-adapter");
  * // create a new SDOAdapter instance, in this case with the latest schema.org vocabulary
  * const mySdoAdapter = await SOA.create({schemaVersion: "latest"});
  * ```