+++ tests/Infrastructure.test.ts
@@ -1,4 +1,4 @@
-import * as Index from "../src/index";
+import { SOA } from "../src/index";
 import {
   isObject,
   isString,
@@ -15,27 +15,27 @@ import { debugFunc, commit } from "./testUtility";
 describe("Infrastructure testing", () => {
   // Check if the retrieval of the versionsFile from schema.org works.
   test("fetchSchemaVersions - file structure", async () => {
-    const schemaVersions = await Index.fetchSchemaVersions(false, commit);
+    const schemaVersions = await SOA.fetchSchemaVersions(false, commit);
     debugFunc(schemaVersions);
     expect(isObject(schemaVersions)).toBe(true);
     expect(isObject(schemaVersions?.releaseLog)).toBe(true);
     expect(isString(schemaVersions?.schemaversion)).toBe(true);
-    expect(isString(await Index.getLatestSchemaVersion(commit))).toBe(true);
+    expect(isString(await SOA.getLatestSchemaVersion(commit))).toBe(true);
   });
 
   // Check if the latest version found in the versionsFile is also the latest valid version elaborated by the schema-org-adapter adapter (schema-org-adapter only marks a version as valid if the corresponding vocabulary file exists)
   test("fetchSchemaVersions - latestVersionIsCorrect", async () => {
-    const schemaVersions = await Index.fetchSchemaVersions(false, commit);
+    const schemaVersions = await SOA.fetchSchemaVersions(false, commit);
     if (!schemaVersions) {
       throw new Error(
-        "Index.fetchSchemaVersions() for commit: " + commit + " not working."
+        "SOA.fetchSchemaVersions() for commit: " + commit + " not working."
       );
     }
     // Sort release entries by the date. latest is first in array
     const sortedVersionsArray = sortReleaseEntriesByDate(
       schemaVersions.releaseLog
     );
-    const latestVersion = await Index.getLatestSchemaVersion(commit);
+    const latestVersion = await SOA.getLatestSchemaVersion(commit);
     // Latest (first) element of Array must be the same as the latest version found by schema-org-adapter
     debugFunc("Latest version by versionsFile: " + sortedVersionsArray[0][0]);
     debugFunc("Latest version by schema-org-adapter: " + latestVersion);
@@ -52,10 +52,10 @@ describe("Infrastructure testing", () => {
     // 2.0 - 3.0 have no jsonld
     // 3.1 - 8.0 have all-layers.jsonld
     // 9.0 + have schemaorg-all-http.jsonld
-    const schemaVersions = await Index.fetchSchemaVersions(false, commit);
+    const schemaVersions = await SOA.fetchSchemaVersions(false, commit);
     if (!schemaVersions) {
       throw new Error(
-        "Index.fetchSchemaVersions() for commit: " + commit + " not working."
+        "SOA.fetchSchemaVersions() for commit: " + commit + " not working."
       );
     }
     for (const currentVersion of Object.keys(schemaVersions.releaseLog)) {
@@ -66,7 +66,7 @@ describe("Infrastructure testing", () => {
         continue;
       }
       // let this function construct the URL. No error should happen.
-      const currentFileURL = await Index.constructURLSchemaVocabulary(
+      const currentFileURL = await SOA.constructURLSchemaVocabulary(
         currentVersion,
         true,
         commit