+++ tests/index.test.ts
@@ -1,4 +1,4 @@
-import * as Index from "../src/index";
+import { SOA } from "../src/index";
 import { commit, debugFuncErr } from "./testUtility";
 import VOC_OBJ_SDO3_7 from "./data/schema-3.7.json";
 import VOC_OBJ_GWON from "./data/graph-with-one-node.json";
@@ -9,7 +9,7 @@ import { getURLBaseSchema, getURLSchemaVersions } from "../lib/utilities";
  */
 describe("SDO Adapter methods", () => {
   test("create() without vocab", async () => {
-    const mySA = await Index.create({
+    const mySA = await SOA.create({
       commit: commit,
       schemaHttps: true,
       onError: debugFuncErr,
@@ -20,7 +20,7 @@ describe("SDO Adapter methods", () => {
   });
 
   test("create() with vocab", async () => {
-    const mySA = await Index.create({
+    const mySA = await SOA.create({
       commit: commit,
       schemaHttps: true,
       onError: debugFuncErr,
@@ -31,24 +31,24 @@ describe("SDO Adapter methods", () => {
   });
 
   test("constructURLSchemaVocabulary()", async () => {
-    const url = await Index.constructURLSchemaVocabulary("9.0", true, commit);
+    const url = await SOA.constructURLSchemaVocabulary("9.0", true, commit);
     expect(url).toBe(
       getURLBaseSchema(commit) + "9.0/schemaorg-all-https.jsonld"
     );
-    const url2 = await Index.constructURLSchemaVocabulary("3.9", false, commit);
+    const url2 = await SOA.constructURLSchemaVocabulary("3.9", false, commit);
     expect(url2).toBe(getURLBaseSchema(commit) + "3.9/all-layers.jsonld");
-    const url3 = await Index.constructURLSchemaVocabulary("9.0", false, commit);
+    const url3 = await SOA.constructURLSchemaVocabulary("9.0", false, commit);
     expect(url3).toBe(
       getURLBaseSchema(commit) + "9.0/schemaorg-all-http.jsonld"
     );
-    const url4 = await Index.constructURLSchemaVocabulary("9.0", true, commit);
+    const url4 = await SOA.constructURLSchemaVocabulary("9.0", true, commit);
     expect(url4).toBe(
       getURLBaseSchema(commit) + "9.0/schemaorg-all-https.jsonld"
     );
-    const url5a = await Index.constructURLSchemaVocabulary("latest");
-    const url5b = await Index.constructURLSchemaVocabulary();
+    const url5a = await SOA.constructURLSchemaVocabulary("latest");
+    const url5b = await SOA.constructURLSchemaVocabulary();
     expect(url5a).toBe(url5b);
-    const url6a = await Index.constructURLSchemaVocabulary(
+    const url6a = await SOA.constructURLSchemaVocabulary(
       "9.0",
       true,
       "9a3ba46"